backend/requirements-base.txt
# Lightweight core dependencies (fast install) for running basic API & email ingest
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
pydantic[email]==2.7.3
python-dotenv==1.0.1
httpx==0.27.0
aiosmtplib==3.0.1
imapclient==3.0.1
email-normalize==2.0.0
textblob==0.18.0.post0
python-multipart==0.0.9
jinja2==3.1.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
